SONARHTML-224 Extract HtmlTable for reuse across table-aware rules
Move the section state, span-aware grid building, and first-row/first-column
queries out of TableHeaderHasIdOrScopeCheck into a new HtmlTable in
org.sonar.plugins.html.api so other table-aware rules can reuse the same
model. The S1102 check shrinks to its visitor wiring plus the
exempt-if-in-first-row-or-first-column predicate; everything else is now
behind HtmlTable's API.
